icon assets and factory
